changed all the sigantures to use IModel2

+++ src/Model/IModel2.java
@@ -15,9 +15,9 @@ public interface IModel2 extends IModel {
 
   void savePortfolio(String portFolioName) throws IOException;
 
-  IPortfolioV2 createPortfolioV2(String ticker, int share, String name, LocalDate date);
+  IPortfolioV2 createPortfolioV2(String ticker, double share, String name, LocalDate date);
 
-  void addToPortfolioV2(String portFolioName, String ticker, int share, LocalDate date);
+  void addToPortfolioV2(String portFolioName, String ticker, double share, LocalDate date);
 
   Double getPortfolioValueV2(String name, LocalDate cal) throws IllegalArgumentException;
 }
